requirements.txt
fastapi==0.115.6
PyMuPDF==1.25.1
Pillow==11.0.0
python-multipart==0.0.12
uvicorn==0.32.1
